Move API code to a separate file, and rename it to sync API
